HDF5Array 1.33.1
Add 'as.vector' argument to h5mread().

Add coercions from CSC_H5SparseMatrixSeed, H5SparseMatrix, and TENxMatrix
to SparseArray. These coercions allow the SparseArray() constructor
function to work out-of-the-box on CSC_H5SparseMatrixSeed,
CSC_H5ADMatrixSeed, TENxMatrixSeed, H5SparseMatrix, and TENxMatrix objects.

+++ src/H5DSetDescriptor.h
@@ -75,7 +75,7 @@ typedef struct h5dset_descriptor_t {
 	hsize_t *h5dim;
 	H5D_layout_t h5layout;
 	hsize_t *h5chunkdim;
-	int *h5nchunk;
+	size_t *h5nchunk;
 } H5DSetDescriptor;
